Small fixes in libfilter to remove some compilation warnings with g++ 4.2.4
git-svn-id: http://dwarf05.bio.dist.unige.it/svn/taos/src/viewer/trunk@21 3f367682-104b-0410-a550-fdff993939f5

+++ src/libfilter/padFilter.cpp
@@ -76,9 +76,9 @@ bool padFilter::Execute( void )
         output.Set(input.Get());
 
         index = input.GetIndex();
-        for (int i = 0; i < Dimension; i++)
+        for (unsigned int i = 0; i < Dimension; i++)
         {
-            if (index[i] < m_PadSize || index[i] > size[i] - 1 - m_PadSize )
+            if (index[i] < m_PadSize || index[i] > (unsigned int)(size[i] - 1 - m_PadSize) )
             {
                 output.Set( m_PadReplaceValue);
                 break;
